Trim trailing whitespace after line break with inline(-block) frame
Move the logic for trimming trailing whitespace after a line break to
the `add_line` method of `BlockFrameDecorator`, so it is always
applied when a new line box is added. This now also trims trailing
whitespace if a text frame is used in block formatting context, but
this may even be desirable.

Also, restore trailing whitespace removed in this way after a split
to address issues such as #2250.